pyproject.toml
[project]
name = "hitalent-test"
version = "0.1.0"
readme = "README.md"
repository = "https://github.com/SuperLalka/hitalent-test.git"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "0.34.0"}
fastapi = "0.115.12"
psycopg2-binary = "2.9.10"
asyncpg = "0.30.0"
SQLAlchemy = "2.0.39"
alembic = "1.15.1"
pydantic-settings = "2.8.1"
pytest = "8.3.5"
pytest-factoryboy = "2.7.0"
Faker = "37.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
